+++ src/user/app.py
@@ -3,9 +3,11 @@ import uuid
 import datetime
 import os
 from flask import Flask, request, jsonify
+from flask_cors import CORS
 from werkzeug.security import generate_password_hash, check_password_hash
 
 app = Flask(__name__)
+CORS(app)
 
 def get_db_connection():
     base_dir = os.path.dirname(os.path.abspath(__file__))